test/distcomp
